Generalize pipeline and add error propegation
The middle stage of the pipeline has been changed to a single
function which accepts a transformation function. This allows the
specific code to be located in the parent function where it is intended
to be used.

Error channels were added to pipeline stages in order to communicate
them to the main goroutine.

An error was corrected in the gen stage that would send a partially full
as well as an empty chunk down the pipeline.

A new error type was added to indicate the condition when decryption is
attempted with the incorrect key.

Documentation has been added for all functions and improved overall